feat: rename phone validation folder
