+++ eox_nelp/urls.py
@@ -31,6 +31,6 @@ urlpatterns = [  # pylint: disable=invalid-name
     path('api/stats/', include('eox_nelp.stats.api.urls', namespace='stats-api')),
     path('stats/', include('eox_nelp.stats.urls', namespace='stats')),
     path('payment-notifications/', include('eox_nelp.payment_notifications.urls', namespace='payment-notifications')),
-    path('api/phone-validation/', include('eox_nelp.phone_validation.api.urls', namespace='phone-validation-api')),
+    path('api/user-profile/', include('eox_nelp.user_profile.api.urls', namespace='user-profile-api')),
     path('api/pearson_vue/', include('eox_nelp.pearson_vue.api.urls', namespace='pearson-vue-api')),
 ]
