+++ eox_nelp/user_profile/api/v1/urls.py
@@ -1,7 +1,7 @@
-"""eox_nelp phone_validation.api v1 urls"""
+"""eox_nelp puser_profile.api v1 urls"""
 from django.urls import path
 
-from eox_nelp.phone_validation.api.v1.views import generate_otp, validate_otp
+from eox_nelp.user_profile.api.v1.views import generate_otp, validate_otp
 
 app_name = "eox_nelp"  # pylint: disable=invalid-name
