+++ eox_nelp/user_profile/frontend/src/components/PhoneValidation/index.scss
